Add sv_currentPak and sv_clientPaks to serverinfo
The former contains the map's pk3 checksum.
The latter contains the checksum of cgame and ui pk3 in case they are in separate pk3s for some reason.

Currently not used as the filesystem will need some reworking to support sv_clientPaks to attempt to use cgame/ui from proper pk3 when not pure.